+++ qsdk/toolboxes/ansatz_generator/uccsd.py
@@ -61,7 +61,7 @@ class UCCSD(Ansatz):
         if var_params == "ones":
             initial_var_params = np.ones((self.n_var_params,), dtype=float)
         elif var_params == "random":
-            initial_var_params = np.random.random((self.n_var_params,), dtype=float)
+            initial_var_params = np.random.random((self.n_var_params,))
         elif var_params == "MP2":
             initial_var_params = self._compute_mp2_params()
         else:
